+++ CHANGELOG.md
@@ -1,5 +1,32 @@
 # Coinbase Go SDK Changelog
 
+## Unreleased
+
+## [0.0.6] - 2024-09-12
+
+### Added
+
+- Return correlation ID from APIError response
+
+## [0.0.5] - 2024-08-29
+
+### Added
+
+- Allow for use with directly set api key
+- Return user facing error type APIError for a server side error
+
+## [0.0.4] - 2024-08-28
+
+### Added
+
+- Add user facing validator statuses
+
+## [0.0.3] - 2024-08-27
+
+### Fixed
+
+- Fixed a bug where we weren't handling the api returned validators properly resulting in an index out of range error.
+
 ## [0.0.2] - 2024-08-27
 
 ### Fixed